src/main/java/co/com/sofka/questions/questionusecases/CreateUseCase.java
package co.com.sofka.questions.questionusecases;

import co.com.sofka.questions.collections.Question;
import co.com.sofka.questions.model.QuestionDTO;
import co.com.sofka.questions.repositories.QuestionRepository;
import org.springframework.stereotype.Service;
import org.springframework.validation.annotation.Validated;
import reactor.core.publisher.Mono;

/**
 * Clase que representa el caso de uso de crear una pregunta.
 * @author Sofka
 * @version 1.0
 * @since 1.0
 * @see QuestionRepository
 * @see QuestionDTO
 * @see Question
 * @see MapperUtils
 */
@Service
@Validated
public class CreateUseCase implements SaveQuestion {
    private final QuestionRepository questionRepository;
    private final MapperUtils mapperUtils;

    public CreateUseCase(MapperUtils mapperUtils, QuestionRepository questionRepository) {
        this.questionRepository = questionRepository;
        this.mapperUtils = mapperUtils;
    }

    /**
     * Metodo que permite crear una pregunta.
     * @param newQuestion DTO de la pregunta a crear.
     * @return Id de la pregunta creada.
     */
    @Override
    public Mono<String> apply(QuestionDTO newQuestion) {
        return questionRepository
                .save(mapperUtils.mapperToQuestion(null).apply(newQuestion))
                .map(Question::getId);
    }
}
